+++ armyknife_src/fargate.py
@@ -5,8 +5,14 @@ import logging
 import boto3
 
 
+def fargate_disabled():
+    if os.getenv('FARGATE_DISABLE'):
+        return True
+    return False
+
+
 def in_fargate():
-    if os.getenv('LAMBDA_TASK_ROOT', None) and not os.getenv('LAMBDA_DISABLE', None):
+    if os.getenv('LAMBDA_TASK_ROOT'):
         return False
     return True
 
@@ -16,7 +22,7 @@ def fork_container(req, actor_id):
     If not running in lambda env, a fargate container is forked off to
     take the entire request and execute with it.
     """
-    if not os.getenv('LAMBDA_TASK_ROOT', None) or os.getenv('LAMBDA_DISABLE', None):
+    if in_fargate() or fargate_disabled():
         return False
     headers = {}
     for k, v in req.headers.items():
@@ -66,6 +72,10 @@ def fork_container(req, actor_id):
                         {
                             'name': 'ACTINGWEB_ACTOR',
                             'value': actor_id
+                        },
+                        {
+                            'name': 'LOG_LEVEL',
+                            'value': 'DEBUG'
                         },
                         {
                             'name': 'APP_HOST_FQDN',